progress week 03092026

+++ src/main/java/com/hospital/hospitalportal/repository/UserRepository.java
@@ -1,7 +1,7 @@
 package com.hospital.hospitalportal.repository;
 
-import com.hospital.hospitalportal.model.User;
+import com.hospital.hospitalportal.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 }
